chore: move host helpers into pkg/helpers
